Update; fix the full width issue on carousel cards
Using the mobile carousel strategy of 'snap points' has a requirement to
use `white-space: nowrap` in order to get things on the same line. This
has the side effect of not actually respecting the width: 100% property
to my expectations, thus the cards are not actually full width. Add in
the margin to zero things out, which aligns it in the middle as well as
add the `white-space: normal` on the child to allow text to wrap.